# optimized count, maximum, minimum in es adapter