… test/unit/adapters/elastic/test_elastic.py